requirements.txt
fastapi[all]
scikit-learn
uvicorn[standard]
google-cloud-storage
gunicorn